Fixed a few things, including eBook landing pages (they need Javascript, it turns out. Gotta eradicate that).
Ported the main search page for ebooks/index.html (used to be PHP). This
does not have "live" counts of languages etc., and they will need to be
inserted as templates using php, for insertion via Jeckyll in the rendered
HTML pages.

Added collapsible.css to the main header template. It's needed for
ebooks/index.html (the search page), and we might want to use it elsewhere.
This provides a "Help" pop-up, and also a collapsible menu, by utilizing
HTML5+CSS.