Update footer.html in 3 ways
Changed `title` attributes compliant with https://www.w3.org/TR/WCAG20-TECHS/H33.html which recommends using title attributes only when there is supplemental information that can be added. The title is now generally set to the title of the destination page, if it is different enough from the link text.

The 'b' is now lowercase in the alt attribute of the ibiblio image.

Height and width attributes added logo to reduce the page rendering time.