Update new_nav.css
Restore spacing that was lost due to mini-reset styling. The impact is difficult to detect due 1) the * selector hits EVERYTHING on the whole site, 2) the DL/DT/DD tags are not searchable in the repo code due to the use of Markdown to generate the HTML. This is the minimum change that will restore the spacing for this one set of tags.